Moved the frontend meta layer onto the proxy seam
no ref

- the meta helpers reached directly into core/server (lib/image's
  blogIcon/cachedImageSizeFromUrl and services/url) instead of going
  through services/proxy, the one sanctioned frontend -> server seam
- this was previously impossible: the proxy -> meta cycle left those
  proxy values undefined at module-eval time inside meta; the prior
  commit broke that cycle, so the meta layer can now consume them
  through the proxy like the theme helpers already do
- exposed cachedImageSizeFromUrl on the proxy alongside the existing
  blogIcon so image-dimensions can drop its direct lib/image require too
- removes five of the frontend's direct core/server requires, leaving the
  meta layer's only remaining server access behind the proxy chokepoint
