+++ ghost/core/core/frontend/meta/asset-url.js
@@ -1,7 +1,7 @@
 const crypto = require('crypto');
 const path = require('path');
 const config = require('../../shared/config');
-const {blogIcon} = require('../../server/lib/image');
+const {blogIcon} = require('../services/proxy');
 const urlUtils = require('../../shared/url-utils');
 const {SafeString} = require('../services/handlebars');
 const assetHash = require('../services/asset-hash');
